tools/rule_manager/requirements.txt
google-auth
requests
python-dotenv
PyYAML
ruamel.yaml